Refactor Nutzap profile composable and signer access
